Fixed issues relating to merge request and added timeZone tests #story[840]

+++ src/main/java/seng302/controllers/Controller.java
@@ -269,7 +269,7 @@ public class Controller implements Initializable, Observer {
     /**
      * displays the current time according to the UTC offset, in the GUI on the overlay
      */
-    public static void setTimeZone(int UTCOffset) {
+    public static void setTimeZone(double UTCOffset) {
         clockString.set(TimeUtils.setTimeZone(UTCOffset));
     }
